refactor: create shared dashboard component
